fix: Improve descriptor extension interoperability
